+++ rooibos.h
@@ -1,5 +1,4 @@
 #ifndef ROOIBOS_H
 #define ROOIBOS_H
 void rooibos(void);
-void rooibos_show_logo(void);
 #endif