+++ splash.h
@@ -1,0 +1,6 @@
+#ifndef SPLASH_H
+#define SPLASH_H
+void splash_start(void);
+void splash_show_logo(void);
+void splash_show_text_at_middle(char *text, int index);
+#endif